Crate edtf

This crate implements the Extended Date/Time Format as of the 2019-02 specification. It contains separate implementations for each level 0, 1, and 2. Notes on the choices made in each level are found in in the module level documentation.

Features:

Lossless. Each parsed Edtf can be formatted again to output exactly the same string.

Strict. Rejects everything the specification rejects as a parse error. All the types make it impossible to construct an invalid Edtf object.

Works with chrono via the optional features = ["chrono"].

General notes on EDTF

It is probably not stated often enough that EDTF and ISO 8601 support only one calendar, the proleptic Gregorian calendar.

The proleptic Gregorian calendar is produced by extending the Gregorian calendar backward to the dates preceding its official introduction in 1582.

Care should be taken whenever writing down dates before its adoption by the country in which a historical record was created. That means 1800s for many countries! You may be unwittingly transcribing a Julian date, or some other calendar, which will be reinterpreted by EDTF as proleptic Gregorian and be a few days-to-weeks off. See the Wikipedia list of adoption dates. [this section needs a link to a good Julian converter].

Level 0 example

use edtf::level_0::Edtf;
let edtf = Edtf::parse("2019-01-07/2020-01").unwrap();
match edtf {
    Edtf::Date(d) => println!("date, {}", d),
    Edtf::Interval(from, to) => {
        println!("interval, {} to {}", from, to);
        println!("(year only: {} to {})", from.year(), to.year());
    }
    Edtf::DateTime(dt) => println!("datetime, {}", dt),
}
// prints:
// interval, 2019-01-07 to 2020-01
// (year only: 2019 to 2020)

Modules

level_0
EDTF Level 0

level_1
EDTF Level 1

Structs

DateComplete
A structure to hold the date portion of a DateTime. It contains a valid date in the proleptic Gregorian calendar.

DateTime
A DateTime object.

Time
The time portion of a DateTime.

Enums

ParseError
The error type for all the parse methods in this crate.

TzOffset
If features = ["chrono"] is enabled, then this can act as a chrono::TimeZone. This can be used to preserve the level of TZ brevity i.e. TzOffset::Hours(_) ends up as +04 instead of +04:00.

Traits

GetTimezone
A helper trait for getting timezone information from some value. (Especially chrono::DateTime or chrono::NaiveDateTime.)