Enum edtf::TzOffset

pub enum TzOffset {
    Utc,
    Hours(i32),
    Minutes(i32),
}

If features = ["chrono"] is enabled, then this can act as a chrono::TimeZone. This can be used to preserve the level of TZ brevity i.e. TzOffset::Hours(_) ends up as +04 instead of +04:00.

Variants

Utc

Z

Hours(i32)

+04 A number of hours only

Minutes(i32)

+04:30 A number of minutes offset from UTC.