Struct edtf::level_1::Date

pub struct Date { /* fields omitted */ }

An EDTF date. Represents a standalone date or one end of a interval.

Implementations

impl Date

Creating a Date

use edtf::level_1::{Date, Certainty};
let _ = Date::from_ymd(2019, 07, 09).and_certainty(Certainty::Uncertain);

pub fn parse(input: &str) -> Result<Self, ParseError>

Parses a Date from a string. Note! This is not part of the EDTF spec. It is merely a convenience, helpful for constructing proper Edtf values programmatically. It does not handle any of the parts of EDTF using two dates separated by a slash, or open/unknown ranges.

use edtf::level_1::Date;
assert_eq!(Date::parse("2019-07-09"), Ok(Date::from_ymd(2019, 07, 09)));
assert_eq!(Date::parse("2019-07"), Ok(Date::from_ymd(2019, 07, 0)));

assert!(Date::parse("2019-07/2020").is_err());

pub fn from_year(year: i32) -> Self

pub fn from_ym(year: i32, month: u32) -> Self

pub fn from_ymd(year: i32, month: u32, day: u32) -> Self

Creates a Date from a year, month and day. If month or day are zero, this is treated as if they have not been specified at all in an EDTF string. So it is invalid to pass month = 0 but day != 0. This function panics on invalid input, including dates that do not exist, like non-leap-year February 29.

use edtf::level_1::Date;
assert_eq!(Date::parse("2019-07-09"), Ok(Date::from_ymd(2019, 07, 09)));
assert_eq!(Date::parse("2019-07"), Ok(Date::from_ymd(2019, 07, 0)));
assert_eq!(Date::parse("2019"), Ok(Date::from_ymd(2019, 0, 0)));

assert!(Date::parse("2019-00-01").is_err());

pub fn from_ymd_opt(year: i32, month: u32, day: u32) -> Option<Self>

Creates a Date from a year, month and day. If month or day are zero, this is treated as if they have not been specified at all in an EDTF string. So it is invalid to pass month=0 but day!=0. This function returns None on invalid input, including dates that do not exist, like non-leap-year February 29.

Month is 1..=12 but can also be a Season as an integer in range 21..=24.

use edtf::level_1::Date;
assert_eq!(Date::parse("2019-07-09"), Ok(Date::from_ymd(2019, 07, 09)));

pub fn certainty(&self) -> Certainty

Returns the Certainty of this date.

pub fn year_in_range(year: i32) -> bool

Checks if a year falls inside the acceptable range of years allowed by this library. This may be a value other than i32::MIN..=i32::MAX. It is currently i32::MIN >> 4 .. i32::MAX >> 4 to allow for a packed representation.

pub fn from_year_masked(year: i32, mask: YearMask) -> Self

Equivalent to 201X or 20XX. The year has its masked digits replaced with zeroes.

Panics if year is out of range.

pub fn from_year_masked_month(year: i32) -> Self

Equivalent to e.g. 2019-XX.

Panics if year is out of range.

pub fn from_year_masked_month_day(year: i32) -> Self

Equivalent to e.g. 2019-XX-XX.

Panics if year is out of range.

pub fn from_ym_masked_day(year: i32, month: u32) -> Self

Equivalent to e.g. 2019-07-XX Panics if year or month is out of range. Month accepts a (1-12) month or (21-24) season.

Panics if year is out of range.

pub fn from_year_season(year: i32, season: Season) -> Self

Equivalent to e.g. 2019-21

Panics if year is out of range.

pub fn and_certainty(&self, certainty: Certainty) -> Self

Returns a new Date with the specified Certainty. This certainty applies to the date as a whole.

pub fn as_precision(&self) -> DatePrecision

Returns an enum more suited to use with a match expression. Note that the DatePrecision type does not contain a date-as-a-whole Certainty field. You can easily fetch it separately with Date::certainty.

The internal representation of Date is packed to reduce its memory footprint, hence this API.

impl Date

pub fn to_chrono(&self) -> Option<NaiveDate>

This is supported on crate feature chrono only.

If this date is complete, i.e. it has a month and a day, produces a chrono::NaiveDate. Also available via an core::convert::TryFrom implementation on chrono::NaiveDate.