Struct edtf::DateTime

pub struct DateTime { /* fields omitted */ }

A DateTime object.

This has minimal introspection methods. It is not an attempt to build a complete DateTime API. Prefer to use its implementation of chrono::Datelike and chrono::Timelike or simply the DateTime::to_chrono method to use a specific chrono::TimeZone, all available with features = ["chrono"].

Implementations

impl DateTime

pub fn date(&self) -> DateComplete

Gets the date portion

pub fn time(&self) -> Time

Gets the time portion

pub fn offset(&self) -> Option<TzOffset>

Get the TzOffset. If None is returned, this represents a timestamp which did not specify a timezone.

If using the chrono interop, None means you should attempt to convert to a chrono::NaiveDate

pub fn to_chrono_naive(&self) -> NaiveDateTime

This is supported on crate feature chrono only.

Convert to a chrono::NaiveDate

pub fn to_chrono<Tz>(&self, tz: &Tz) -> DateTime<Tz> where
Tz: TimeZone,

This is supported on crate feature chrono only.

use edtf::level_1::Edtf;
use chrono::TimeZone;

let utc = chrono::Utc;
assert_eq!(
    Edtf::parse("2004-02-29T01:47:00+05:00")
        .unwrap()
        .as_datetime()
        .unwrap()
        .to_chrono(&utc),
    utc.ymd(2004, 02, 28).and_hms(20, 47, 00)
);