Trait edtf::GetTimezone

pub trait GetTimezone {
    fn tz_offset(&self) -> Option<TzOffset>;
}

A helper trait for getting timezone information from some value. (Especially chrono::DateTime or chrono::NaiveDateTime.)

Implementations for the chrono types are included with feature = ["chrono"].

Not implemented on DateTime because this is only used as a bound on impl<T> From<T> for DateTime implementations.

Required methods

fn tz_offset(&self) -> Option<TzOffset>

Return the number of seconds difference from UTC.

None represents NO timezone information on the EDTF timestamp.

Some(TzOffset::Utc) represents a Z timezone, i.e. UTC/Zulu time.

Some(TzOffset::Hours(1)) represents `+01

Some(TzOffset::Minutes(-16_200)) represents -04:30

Loading content...